requirements.txt
streamlit>=1.32.0
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.118.0
bcrypt>=4.1.2
